fix: corrigindo campos/atributos

+++ Projeto/backend/client/src/main/java/com/localfarm/client/application/services/ClientService.java
@@ -2,13 +2,11 @@ package com.localfarm.client.application.services;
 
 import com.localfarm.client.domain.models.Client;
 import com.localfarm.client.domain.models.repositories.ClientRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
-
-import jakarta.persistence.EntityNotFoundException;
 
 import java.util.Optional;
 
@@ -23,6 +21,7 @@ public class ClientService {
     }
 
     public Client createClient(Client client) {
+        validateClient(client);
         return clientRepository.save(client);
     }
 
@@ -30,12 +29,16 @@ public class ClientService {
         Optional<Client> existingClient = clientRepository.findById(id);
         if (existingClient.isPresent()) {
             client.setId(id);
+            validateClient(client);
             return clientRepository.save(client);
         }
         throw new EntityNotFoundException("Client not found");
     }
 
     public void deleteClient(Long id) {
+        if (!clientRepository.existsById(id)) {
+            throw new EntityNotFoundException("Client not found");
+        }
         clientRepository.deleteById(id);
     }
 
@@ -45,5 +48,15 @@ public class ClientService {
 
     public Optional<Client> getClientById(Long id) {
         return clientRepository.findById(id);
+    }
+
+    private void validateClient(Client client) {
+        if (client.getType() == null) {
+            throw new IllegalArgumentException("Client type is mandatory");
+        }
+
+        if (client.getCode() == null || client.getCode().trim().isEmpty()) {
+            throw new IllegalArgumentException("Client code is mandatory");
+        }
     }
 }
